preparing second query. the new map is not working

+++ api/src/main/java/ar/edu/itba/pod/tp2/Reducers/CountReducerFactory.java
@@ -20,7 +20,7 @@ public class CountReducerFactory implements ReducerFactory<String, Integer, Inte
 
         @Override
         public void reduce(Integer number) {
-            count += number.intValue();
+            count += number;
         }
 
         @Override
